requirements-dev.txt
pillow>=9.0.0
pillow-heif>=0.10.0
tqdm>=4.62.0
psutil>=5.9.0
pytest>=7.0.0
piexif>=1.1.3
pytest-cov>=2.12.1